Fix member list population
In cases where the member is not created via the UI, the metadata used
for filtering the members on retrieval is not added to the VMs. Hence any
member VMs that are created outside the UI are never shown in the member list.

This patch fixes this issue by reinstating the earlier filtering logic to
find the member VMs belonging to a particular PTG in case the above metadata
is not present.

Change-Id: Ic2154d04c292269dd7cb46edf19f3a7f3bbe1364
Closes-bug: 1691598