Restore network/port options for create instance dialog
An earlier commit:
https://github.com/openstack/group-based-policy-ui/commit/d86c6ea3cc8dc402c1584d541710a268d7c54ac0

patched the "create instance" dialog such that even when
its launched from Project -> Compute -> Instances, the "Groups" workflow step
is shown instead of the expected "Networks" and "Ports" steps. The former
should be shown only in the case when the dialog is launched from GBP and
the original behavior should be preserved when the dialog is launched from
Compute. This is being fixed here.

Change-Id: Ie9af2cca451942d51ea254bdf14bc2345743f938