Fix patching of create instance dialog (part deux)
Patching of the create instance dialog needs to happen
only when invoked from the GBP "create member" workflow.
An earlier commit c0f6597c7f94811e94f8233b708401c24d6edf06
fixed this for launching from the Compute menu but left out
the launching option from the Network Topology (under
the Network option).

An attempt was made to change the logic in a way that the check
happens for the GBP URI (instead of the logic to check for all
other URIs from where it can be launched), however that did not
work. Hence the comment on evolving to a more resrtictive logic
has been preserved in the code.

Also fixes tox.ini

Change-Id: If124971c1954cb85fce244440e2408cd58ced3e6