Fix patching of create instance dialog (part trois)
Previous patch fixed the form but did not fix yet
another bug in the original implementation of the
patching of the create instance workflow. Namely,
the patched logic was always assuming the need for
a PTG, and would override the supplied network
arguments. This is being fixed here.

Also caps the cmd2 dependency whose latest version is
incompatible.

Change-Id: If0fff67cd2b811853a632055d1609b6057769bb8